nhanvat va xu

+++ core/src/main/java/farm/com/GameState.java
@@ -1,7 +1,7 @@
 package farm.com;
 
 public class GameState {
-    public static float money = 100;
-    public static float meatTotal = 0;
+    public static int money = 100;
+    public static int meatTotal = 0;
 
 }
